waypoints, camera, load pcs for mission, offset in hub, general changes
waypoint movement- the movement itself needs to be changed/fixed but it now uses a waypoint system. it will also be changed so that it does not always use waypoints but rather it is for the player to decide

camera- added zoom function and a edgescroll function to the mission camera.

load pcs for missions- the pcs ou recruit are now loaded (somewhat) into the mission setting OBS! does not work properly but better than nothing.

offset in hub- new offsets, the players should show up in the right places unless you click marcus' button

general changes- created several scripts used for things, ask me if they don't make sense

OBS! if you start in mission view there will not abe any player character, start in the hub and get some characters from the leftmost building before going on mission